+++ CHANGES.md
@@ -1,6 +1,7 @@
 ## 1.1.0 (PENDING)
 
 * Add support for streaming reads.
+* Implement `Blobby.store("s3://bucket/prefix")`
 
 ## 1.0.0 (2015-10-02)